fix: timetable data fetching 로직 위로 올리기
