+++ src/pages/TimetablePage/index.tsx
@@ -8,12 +8,10 @@ import MyLectureTimetable from 'components/TimetablePage/DefaultPage/MyLectureTi
 import MyLectureList from 'components/TimetablePage/DefaultPage/MyLectureList';
 import Curriculum from 'components/TimetablePage/DefaultPage/Curriculum';
 import { useRecoilValue } from 'recoil';
-import { myLecturesAtom, selectedSemesterAtom, selectedTempLectureSelector } from 'utils/recoil/semester';
+import { myLecturesAtom, selectedSemesterAtom } from 'utils/recoil/semester';
 import useTokenState from 'utils/hooks/useTokenState';
 import useTimetableInfoList from 'components/TimetablePage/hooks/useTimetableInfoList';
 import useTimetableDayList from 'utils/hooks/useTimetableDayList';
-import useLectureList from 'components/TimetablePage/hooks/useLectureList';
-import { LectureInfo } from 'interfaces/Lecture';
 import styles from './TimetablePage.module.scss';
 
 function TimetablePage() {
@@ -31,15 +29,6 @@ function TimetablePage() {
       : (myLecturesFromLocalStorageValue ?? []),
   );
 
-  const selectedLecture = useRecoilValue(selectedTempLectureSelector);
-  const { data: lectureList } = useLectureList(selectedSemester);
-  const similarSelectedLecture = (lectureList as unknown as Array<LectureInfo>)
-    ?.filter((lecture) => lecture.code === selectedLecture?.code)
-    ?? [];
-  const similarSelectedLectureDayList = useTimetableDayList(similarSelectedLecture);
-  const selectedLectureIndex = similarSelectedLecture
-    .findIndex(({ lecture_class }) => lecture_class === selectedLecture?.lecture_class);
-
   return (
     <div className={styles.page}>
       {!isMobile ? (
@@ -49,11 +38,7 @@ function TimetablePage() {
             {/* 강의 목록 */}
             <LectureList />
             {/* 나의 시간표 타임 테이블 */}
-            <MyLectureTimetable
-              lectures={myLectureDayValue}
-              similarSelectedLecture={similarSelectedLectureDayList}
-              selectedLectureIndex={selectedLectureIndex}
-            />
+            <MyLectureTimetable lectures={myLectureDayValue} />
             {/* 나의 시간표 강의 목록 */}
             <MyLectureList />
             {/* 시간표 커리큘럼 */}
@@ -61,11 +46,7 @@ function TimetablePage() {
           </div>
         </>
       ) : (
-        <MobilePage
-          lectures={myLectureDayValue}
-          similarSelectedLecture={similarSelectedLectureDayList}
-          selectedLectureIndex={selectedLectureIndex}
-        />
+        <MobilePage lectures={myLectureDayValue} />
       ) }
     </div>
   );
